use custom hooks for socket stuff
