+++ frontend/src/components/Room/hooks/index.ts
@@ -1,5 +1,6 @@
 import usePageNum from "./usePageNum"
 import usePointer from "./usePointer"
+import useSocket from "./useSocket"
 import useZoom from "./useZoom"
 
-export { usePageNum, usePointer, useZoom }
+export { usePageNum, usePointer, useSocket, useZoom }
